feat: 补齐器件级求解器最后拼图（2D EME + 统一调度器）
新增 2D 任意截面 EME 求解器（eme_2d.py）和统一器件级求解器调度器
（device_solver.py），覆盖 auto/eme/fde/rcwa/varfdtd/bpm/fdtd 七种方法。

- eme_2d.py: mode_overlap_2d + solve_eme_2d，公共网格重采样 + Redheffer 星积级联
- device_solver.py: DeviceSolverConfig/Result + DeviceLevelSolver + solve_device
- 自动选择策略: 1D slab→EME / 2D 截面→EME 2D / 周期→RCWA / 2.5D→varFDTD / 全 3D→FDTD
- 测试: test_eme_2d.py(18) + test_device_solver.py(22) = 40 passed in 7.21s
- R05 修复: mode_overlap_2d ComplexWarning（float→complex 保留虚部）

文献来源（R02）: Bienstman 2001 / Smit 1996 / Snyder&Love 1983 / Collin 2001 等
规则依据: R02/R03/R04/R05/R11，纯 numpy/scipy CPU 实现，无 fall-back

+++ modules/eme/src/polaris_eme/eme_2d.py
@@ -105,7 +105,7 @@ def mode_overlap_2d(
         raise ValueError(f"dx_um 须 > 0，得到 {dx_um}")
     if dy_um <= 0:
         raise ValueError(f"dy_um 须 > 0，得到 {dy_um}")
-    overlap = float(np.sum(fa * np.conj(fb)) * dx_um * dy_um)
+    overlap = np.sum(fa * np.conj(fb)) * dx_um * dy_um
     return complex(overlap)
